# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Urja Meter API",
    "description": "\nA clean, read-only REST API over the **Urja Meter Ops** portal (smart-meter data for a\nJaipur distribution utility). The portal has no API of its own. This service logs in as a\nnormal user, reads the portal's internal endpoints (see `PROTOCOL.md` in the repository)\nand serves the data as typed JSON.\n\n* **Meters**: nameplate, status, location, network position; filter by any attribute,\n  network node or distance.\n* **Readings / consumption**: register readings normalised to numbers and ISO timestamps\n  (IST), with per-interval and per-day consumption derived from register deltas.\n* **Network**: the zone \u2192 circle \u2192 division \u2192 subdivision \u2192 substation \u2192 feeder \u2192\n  transformer hierarchy, reconstructed and repaired from per-meter data.\n* **Data quality**: every value we repaired or could not reconcile, with the portal's\n  original value.\n\nReference data is served from a local index synced from the portal's bulk export;\nreadings are fetched on demand and cached. `GET /v1/status` shows how fresh everything is.\nErrors use `application/problem+json` (RFC 9457) with a stable `code` (listed under the\n`Problem` schema). Unknown query parameters are rejected rather than ignored.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/status": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Sync freshness and portal client health",
        "operationId": "status",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/sync": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Re-sync reference data from the portal now",
        "description": "Joins a sync that is already running, and returns the last result instead of\nsyncing again if one succeeded less than 30 s ago, so repeated calls can't spend the\nportal's rate budget.",
        "operationId": "trigger_sync",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
    "/v1/transformers": {
      "get": {
        "tags": [
          "transformers"
        ],
        "summary": "List distribution transformers",
        "operationId": "list_transformers",
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "feeder",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only transformers on this feeder, e.g. `F-001`.",
              "title": "Feeder"
            },
            "description": "Only transformers on this feeder, e.g. `F-001`."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Page size.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Page size."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Items to skip.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Items to skip."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Transformer_"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transformers/{code}": {
      "get": {
        "tags": [
          "transformers"
        ],
        "summary": "Get one transformer with its meters",
        "operationId": "get_transformer",
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Transformer code (any case).",
              "examples": [
                "DT-001"
              ],
              "title": "Code"
            },
            "description": "Transformer code (any case)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransformerDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "properties": {
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location",
            "description": "Where the value came from, e.g. `[\"query\", \"from\"]`."
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Machine-readable kind, e.g. `value_error` or `unknown_parameter`."
          }
        },
        "type": "object",
        "required": [
          "location",
          "message",
          "type"
        ],
        "title": "FieldError"
      },
...
      "MeterStatus": {
        "type": "string",
        "enum": [
          "installed",
          "faulty",
          "decommissioned",
          "unknown"
        ],
        "title": "MeterStatus"
      },
...
      "NetworkNodeRef": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "DT-001"
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "examples": [
              "Malviya Nagar DT 1"
            ]
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "NetworkNodeRef"
      },
...
      "NetworkPath": {
        "properties": {
          "zone": {
            "$ref": "#/components/schemas/NetworkNodeRef"
          },
          "circle": {
            "$ref": "#/components/schemas/NetworkNodeRef"
          },
          "division": {
            "$ref": "#/components/schemas/NetworkNodeRef"
          },
          "subdivision": {
            "$ref": "#/components/schemas/NetworkNodeRef"
          },
          "substation": {
            "$ref": "#/components/schemas/NetworkNodeRef"
          },
          "feeder": {
            "$ref": "#/components/schemas/NetworkNodeRef"
          },
          "transformer": {
            "$ref": "#/components/schemas/NetworkNodeRef"
          }
        },
        "type": "object",
        "required": [
          "zone",
          "circle",
          "division",
          "subdivision",
          "substation",
          "feeder",
          "transformer"
        ],
        "title": "NetworkPath",
        "description": "Where a meter sits in the network, from zone down to its distribution transformer."
      },
...
      "Page_Transformer_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Transformer"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Number of items matching the query (ignoring limit/offset)."
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[Transformer]"
      },
...
      "PortalClientStatus": {
        "properties": {
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "logins": {
            "type": "integer",
            "title": "Logins"
          },
          "requests": {
            "type": "integer",
            "title": "Requests"
          },
          "retries": {
            "type": "integer",
            "title": "Retries"
          },
          "rate_limited": {
            "type": "integer",
            "title": "Rate Limited",
            "description": "429 responses received from the portal."
          },
          "upstream_errors": {
            "type": "integer",
            "title": "Upstream Errors",
            "description": "Network errors and 5xx responses (before retries)."
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_success_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success At"
          },
          "session_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Expires At"
          },
          "clock_skew_s": {
            "type": "number",
            "title": "Clock Skew S",
            "description": "Portal clock minus ours, from its Date header (used for signing)."
          }
        },
        "type": "object",
        "required": [
          "base_url",
          "logins",
          "requests",
          "retries",
          "rate_limited",
          "upstream_errors",
          "last_error",
          "last_success_at",
          "session_expires_at",
          "clock_skew_s"
        ],
        "title": "PortalClientStatus"
      },
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code:\n\n* `unauthorized` (401): `X-API-Key` is missing or wrong (only when the server sets `URJA_API_KEY`).\n* `meter_not_found` (404): No meter with that id in the index, or the portal no longer has it.\n* `transformer_not_found` (404): No distribution transformer with that code.\n* `network_node_not_found` (404): No network node with that level and code.\n* `not_found` (404): No such path.\n* `method_not_allowed` (405): The path exists, but not for this HTTP method.\n* `http_error` (4xx): Any other HTTP-level error.\n* `validation_error` (422): A parameter is malformed, out of range or not recognised; `errors` says which.\n* `granularity_unavailable` (422): Hourly consumption asked of a meter that reports once a day.\n* `internal_error` (500): A bug in this service; the server log has the details.\n* `upstream_auth_failed` (502): This service could not log in to the portal.\n* `upstream_protocol_error` (502): The portal answered in a format this service does not understand.\n* `sync_rejected` (502): A sync returned data that looked broken; the previous snapshot is kept.\n* `index_not_ready` (503): The first sync with the portal hasn't finished; `Retry-After` gives the next attempt.\n* `upstream_unavailable` (503): The portal is down or too slow, and there is no cached copy to serve.\n* `upstream_rate_limited` (503): The portal kept answering 429, and there is no cached copy to serve.\n* `busy` (503): Too many requests are queued for the portal's rate budget.",
            "examples": [
              "meter_not_found"
            ]
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Field-level validation errors."
          }
        },
        "type": "object",
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "description": "RFC 9457 problem details."
      },
...
      "ReadingsCacheStatus": {
        "properties": {
          "meter_count": {
            "type": "integer",
            "title": "Meter Count",
            "description": "Meters whose readings are cached."
          },
          "meters_total": {
            "type": "integer",
            "title": "Meters Total",
            "description": "Meters in the index (what a complete cache would hold)."
          },
          "oldest_fetch": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Fetch"
          },
          "newest_fetch": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Newest Fetch"
          },
          "first_reading_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Reading At"
          },
          "last_reading_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reading At",
            "description": "Latest reading held for any meter."
          },
          "ttl_s": {
            "type": "integer",
            "title": "Ttl S"
          },
          "warmup": {
            "$ref": "#/components/schemas/WarmupStatus"
          }
        },
        "type": "object",
        "required": [
          "meter_count",
          "meters_total",
          "oldest_fetch",
          "newest_fetch",
          "first_reading_at",
          "last_reading_at",
          "ttl_s",
          "warmup"
        ],
        "title": "ReadingsCacheStatus"
      },
...
      "ReferenceDataStatus": {
        "properties": {
          "last_success_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success At"
          },
          "age_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age S",
            "description": "Seconds since the last successful sync."
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "description": "True if no sync has succeeded yet, or the last successful one is older than two sync intervals."
          },
          "source": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "export",
                  "crawl"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "interval_s": {
            "type": "integer",
            "title": "Interval S"
          },
          "last_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SyncRun"
              },
              {
                "type": "null"
              }
            ],
            "description": "The most recent sync attempt, successful or not."
          }
        },
        "type": "object",
        "required": [
          "last_success_at",
          "age_s",
          "stale",
          "source",
          "interval_s",
          "last_run"
        ],
        "title": "ReferenceDataStatus"
      },
      "Status": {
        "properties": {
          "index_ready": {
            "type": "boolean",
            "title": "Index Ready",
            "description": "True once reference data has been synced at least once."
          },
          "reference_data": {
            "$ref": "#/components/schemas/ReferenceDataStatus"
          },
          "readings_cache": {
            "$ref": "#/components/schemas/ReadingsCacheStatus"
          },
          "portal": {
            "$ref": "#/components/schemas/PortalClientStatus"
          }
        },
        "type": "object",
        "required": [
          "index_ready",
          "reference_data",
          "readings_cache",
          "portal"
        ],
        "title": "Status"
      },
      "SyncResult": {
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "export",
              "crawl"
            ],
            "title": "Source"
          },
          "meter_count": {
            "type": "integer",
            "title": "Meter Count"
          },
          "transformer_count": {
            "type": "integer",
            "title": "Transformer Count",
            "description": "Transformers served: those with at least one meter."
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "source",
          "meter_count",
          "transformer_count",
          "warnings"
        ],
        "title": "SyncResult"
      },
...
      "TransformerDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "DT-001"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Malviya Nagar DT 1"
            ]
          },
          "feeder_code": {
            "type": "string",
            "title": "Feeder Code",
            "examples": [
              "F-001"
            ]
          },
          "capacity_kva": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Kva",
            "examples": [
              100
            ]
          },
          "network": {
            "$ref": "#/components/schemas/NetworkPath",
            "description": "The transformer's own position in the network (zone \u2192 feeder)."
          },
          "meter_count": {
            "type": "integer",
            "title": "Meter Count"
          },
          "meters_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/MeterStatus"
            },
            "type": "object",
            "title": "Meters By Status"
          },
          "name_variants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Name Variants",
            "description": "Other names meters report for this transformer (stale data)."
          },
          "meters": {
            "items": {
              "$ref": "#/components/schemas/MeterSummary"
            },
            "type": "array",
            "title": "Meters",
            "description": "Meters connected to this transformer."
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "feeder_code",
          "network",
          "meter_count",
          "meters_by_status",
          "meters"
        ],
        "title": "TransformerDetail"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "Required only if the server is configured with URJA_API_KEY.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}